[FVM][TEST] Splitting fvm_test.go
